+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   protocols such as `file://` or `gopher://`
 - `Restify::Request#uri`, now always parsed and returned as `Addressable::URI`
 - Drop EOL Ruby versions 3.1 and 3.2
+- Drop `hitimes` dependency, using `Process.clock_gettime` for timeouts
 
 ### Fixes